refactor: page components
