+++ src/components/Landing.tsx
@@ -2,7 +2,6 @@ import { FormEvent, useState } from "react";
 import { useNavigate } from "@tanstack/react-router";
 import { useTrialContext } from "@/contexts/TrialContext";
 
-
 export function Landing() {
   const [name, setName] = useState("");
   const [error, setError] = useState("");
@@ -18,39 +17,34 @@ export function Landing() {
     }
 
     dispatch({ type: "started", name: name.trim() });
-    navigate({to: "/task"})
+    navigate({ to: "/task" });
   }
 
   return (
-    <main className="grid min-h-svh place-content-center">
-      <form
-        onSubmit={handleSubmit}
-        className="flex flex-col gap-4 items-center"
+    <form onSubmit={handleSubmit} className="flex flex-col gap-6 items-center text-2xl">
+      <label htmlFor="name">Enter your name:</label>
+      <input
+        id="name"
+        placeholder="Name"
+        value={name}
+        className="text-center border-b-2 p-2 focus:outline-none focus:ring-2 focus:ring-indigo-500"
+        onChange={(event) => {
+          setName(event?.currentTarget.value);
+          if (error) setError("");
+        }}
+        aria-required="true"
+      />
+      {error && (
+        <p role="alert" className="text-red-500">
+          {error}
+        </p>
+      )}
+      <button
+        type="submit"
+        className="border-2 p-2 hover:cursor-pointer hover:bg-indigo-100 focus:outline-none focus:ring-2 focus:ring-indigo-500"
       >
-        <label htmlFor="name">Enter your name:</label>
-        <input
-          id="name"
-          placeholder="Name"
-          value={name}
-          className="text-center border-b-2 p-2 focus:outline-none focus:ring-2 focus:ring-indigo-500"
-          onChange={(event) => {
-            setName(event?.currentTarget.value);
-            if (error) setError("");
-          }}
-          aria-required="true"
-        />
-        {error && (
-          <p role="alert" className="text-red-500">
-            {error}
-          </p>
-        )}
-        <button
-          type="submit"
-          className="border-2 p-2 hover:cursor-pointer hover:bg-indigo-100 focus:outline-none focus:ring-2 focus:ring-indigo-500"
-        >
-          Start Game
-        </button>
-      </form>
-    </main>
+        Start Game
+      </button>
+    </form>
   );
 }
